// src/enums/IframeSyncEnum.ts
export enum IframeSyncEnum {
  IFRAME_READY = 'iframeReady',
  MOBILE_READY = 'mobileReady',

  REPLACE_PATH = 'replacePath',
  UPDATE_THEME = 'updateTheme',
  UPDATE_LOCALE = 'updateLocale',

  CREATE_BACK_EL = 'createBackEl',
  GET_VAR = 'getVar',
  SET_VAR = 'setVar',
  SET_VAR_V2 = 'setVarV2'
}